add way to init the system updates
When you just install the system updates, the system update table will
be initialized on the current version of the code you have. But
sometimes you might to init on an older build version since you have
just installed the plugin but did not yet have all the updates in your
database.

Signed-off-by: Ike Devolder <ike.devolder@studioemma.eu>